Add MIT License to project
This commit adds the MIT License to the project, granting users permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software. The license also disclaims all warranties of merchantability, fitness for a particular purpose, and noninfringement, and limits the liability of the authors and copyright holders for any claim, damages or other liability arising from the software or the use or other dealings in the software.